% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci-package.R
\docType{package}
\name{aci-package}
\alias{aci-package}
\title{aci: Assimilative Causal Inference for Stochastic Dynamical Systems}
\description{
Causal inference as a Bayesian inverse problem via data assimilation. Implements the assimilative causal inference (ACI) metric, forward and backward causal influence ranges (CIR), closed-form conditional-Gaussian (CGNS) filtering/smoothing with a lagged-divergence table, ensemble Kalman-Bucy filter/smoother, and a collection of benchmark models. Based on Andreou, Chen & Bollt (2026) \doi{10.1038/s41467-026-68568-0}, Andreou & Chen (2026) \doi{10.48550/arXiv.2510.21889}, Jiang, Andreou, Reich & Chen (2026) \doi{10.48550/arXiv.2604.25157}, Moser, Chen & Andreou (2026) \doi{10.48550/arXiv.2605.22692}, and the adaptive online smoother of Andreou, Chen & Li (2026) \doi{10.1007/s00332-026-10271-x}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aidanmoller/aci}
  \item Report bugs at \url{https://github.com/aidanmoller/aci/issues}
}

}
\author{
\strong{Maintainer}: Aidan Moller \email{aidan.moller@adelaide.edu.au} [copyright holder]

Other contributors:
\itemize{
  \item Marios Andreou (Author and copyright holder of the MIT-licensed reference MATLAB implementations 'ACI_code' and 'FBCIR_code', from which this package's numerical core is translated) [copyright holder]
  \item Zhang Jiang (Author and copyright holder of the MIT-licensed reference MATLAB implementation 'EnKBS', from which the golden test port of the ensemble dyad experiment is translated) [copyright holder]
}

}
\keyword{internal}
