% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActuarialClimateIndex.R
\docType{class}
\name{ActuarialClimateIndex-class}
\alias{ActuarialClimateIndex-class}
\alias{ActuarialClimateIndex}
\title{ActuarialClimateIndex Class}
\arguments{
\item{temperature_data_path}{Path to the temperature data file}

\item{precipitation_data_path}{Path to the precipitation data file}

\item{wind_u10_data_path}{Path to the U component wind data file}

\item{wind_v10_data_path}{Path to the V component wind data file}

\item{country_abbrev}{Country abbreviation for sea level data}

\item{mask_data_path}{Path to the geographical mask file}

\item{study_period}{Study period for analysis (vector of two dates)}

\item{reference_period}{Reference period for standardization (vector of two dates)}

\item{factor}{Adjustment factor for sea level}
}
\value{
A new ActuarialClimateIndex object

Calculate Actuarial Climate Index

This method calculates the Actuarial Climate Index (ACI) using the initialized climate components.

A data.table containing the calculated values of the Actuarial Climate Index (ACI)
}
\description{
This class represents the Actuarial Climate Index, which is calculated from several climate components such as temperature, precipitation, drought, wind, and sea level.
}
\section{Fields}{

\describe{
\item{\code{temperature_component}}{Temperature component of type ANY}

\item{\code{precipitation_component}}{Precipitation component of type ANY}

\item{\code{drought_component}}{Drought component of type ANY}

\item{\code{wind_component}}{Wind component of type ANY}

\item{\code{sealevel_component}}{Sea level component of type ANY}

\item{\code{study_period}}{Study period for analysis (vector of two dates)}

\item{\code{reference_period}}{Reference period for standardization (vector of two dates)
Initialize ActuarialClimateIndex object

This method initializes an object of the ActuarialClimateIndex class with the necessary data file paths and study/reference periods.}
}}


