% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_edge_noise.R
\name{add_edge_noise}
\alias{add_edge_noise}
\title{Add a random noise to edges}
\usage{
add_edge_noise(
  mat = mat,
  noise_type = "add.or.delete",
  n_noise_edge = NULL,
  noise_prop = 0.05,
  directed = TRUE,
  FUN_rounding = round,
  seed = NULL
)
}
\arguments{
\item{mat}{an adjacency matrix.}

\item{noise_type}{a character: "add", "delete", "rewire", or "add.or.delete" (default)}

\item{n_noise_edge}{a integer that indicates the number of edges to be "noised". Detaulf is NULL.}

\item{noise_prop}{proportion of the noised edge calculated from the imput matrix.}

\item{directed}{logical indicating if the graph is treated as directed network (default is TRUE)}

\item{FUN_rounding}{Function for the rounding of the number of n_noise_edge. Detaulf is "round", but it can be replaced with other functions.}

\item{seed}{an integer for the seed (Default is NULL).}
}
\value{
an adjacency matrix with some noise of edges added.
}
\description{
Add a random noise to edges
}
