% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.obs_traj}
\alias{as.data.frame.obs_traj}
\title{Coerce an observed trajectory to a data frame}
\usage{
\method{as.data.frame}{obs_traj}(x, ...)
}
\arguments{
\item{x}{An \code{obs_traj} object.}

\item{...}{Ignored, for consistency with the generic.}
}
\value{
A data frame in long form with columns \code{t}, \code{var} and \code{value}.
}
\description{
Coerce an observed trajectory to a data frame
}
