% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avedig.dens.switch.R
\name{avedig.dens.switch}
\alias{avedig.dens.switch}
\title{Calculating average degree from density, vise versa}
\usage{
avedig.dens.switch(N, Density = NULL, Ave.Degree = NULL, mode = "digraph")
}
\arguments{
\item{N}{The number of the network nodes.}

\item{Density}{Density of the network.}

\item{Ave.Degree}{Average degree of the network.}

\item{mode}{Mode of the network object, Default = \code{digraph}.}
}
\description{
Calculating average degree from density, vise versa
}
