% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{cgns_library}
\alias{cgns_library}
\title{Polynomial dictionary over the full state z = (x, y) with CGNS tags.}
\usage{
cgns_library(k, l, degree = 2, intercept = TRUE)
}
\arguments{
\item{k}{Number of observed variables.}

\item{l}{Number of hidden variables.}

\item{degree}{Maximum polynomial degree.}

\item{intercept}{Whether to include the constant term.}
}
\description{
Polynomial dictionary over the full state z = (x, y) with CGNS tags.
}
\examples{
cgns_library(k = 1, l = 1, degree = 2)

}
