% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeaLevelData.R
\name{copy_and_rename_files_by_country}
\alias{copy_and_rename_files_by_country}
\title{Copier et renommer les fichiers en fonction de l'abreviation du pays}
\usage{
copy_and_rename_files_by_country(abbreviation, source_dir, df)
}
\arguments{
\item{abbreviation}{Abreviation du pays pour filtrer et copier les fichiers associes}

\item{source_dir}{Repertoire source contenant les fichiers .rlrdata}

\item{df}{DataFrame contenant les donnees PSMSL pour filtrer par abreviation de pays}
}
\value{
NULL Renvoie NULL si aucune entree n'est trouvee pour l'abreviation du pays
}
\description{
Copie les fichiers .rlrdata d'un repertoire source vers un repertoire cible base
sur l'abreviation du pays fournie. Les fichiers copies sont ensuite renommes en .txt.
}
\examples{
# Execution de la fonction pour une abreviation specifique
# Repertoire contenant les fichiers .rlrdata
#psmsl <- load_PSMSLdata('./inst/externaldata/psmsl_data.csv')
data(psmsl)
source_dir <- './inst/externaldata/rlr_monthly/data/'
copy_and_rename_files_by_country("FRA", source_dir, psmsl)

}
