% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.network.R
\name{create.network}
\alias{create.network}
\title{Simulation: Create Simulating Networks}
\usage{
create.network(N, Belief, Degree)
}
\arguments{
\item{N}{The number of nodes.}

\item{Belief}{The belief scores of each nodes.}

\item{Degree}{The target degree.}
}
\value{
\code{time}: time used for creating the network,
        \code{mat}: the imported network matrix,
        \code{Belief}: the belief scores of each nodes,
        \code{Degree}: the target degree.
}
\description{
Network simulation based on agent-based models (ABM).
}
