% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeaLevelData.R
\docType{data}
\name{csv_path}
\alias{csv_path}
\title{Charger le fichier CSV contenant les données PSMSL}
\format{
An object of class \code{character} of length 1.
}
\usage{
csv_path
}
\arguments{
\item{csv_path}{Chemin vers le fichier CSV à charger}
}
\value{
Affiche un message indiquant si le DataFrame a été chargé avec succès
ou si une erreur s'est produite
}
\description{
Vérifie si un fichier CSV spécifique existe dans le répertoire de données.
Si le fichier existe, il est chargé dans un DataFrame. Sinon, une erreur est levée.
}
\examples{
# Chemin du fichier CSV
csv_path <- '../data/required_data/psmsl_data.csv'
# Charger le fichier CSV
# (Voir le code complet ci-dessous pour l'exécution)

}
\keyword{datasets}
