% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeaLevelData.R
\name{download_monthlyRLRdata}
\alias{download_monthlyRLRdata}
\title{Telecharge et decompresse les donnees RLR mensuelles}
\usage{
download_monthlyRLRdata(url, destination_dir)
}
\arguments{
\item{url}{URL du fichier a telecharger.}

\item{destination_dir}{Chemin vers le repertoire de destination.}
}
\value{
Affiche un message indiquant si le telechargement et la decompression sont termines,
ou si le dossier existe deja.
}
\description{
Telecharge un fichier ZIP contenant les donnees RLR mensuelles depuis une URL specifiee,
le decompresse dans un repertoire de destination, et supprime le fichier ZIP apres l'extraction.
Si le dossier cible existe deja, aucune action n'est effectuee.
}
\examples{
# URL du fichier a telecharger:
#url <- "https://psmsl.org/data/obtaining/rlr.monthly.data/rlr_monthly.zip"
#destination_dir <- "./inst/externaldata/"
#download_monthlyRLRdata(url, destination_dir)

}
