% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCopernicusData.R
\docType{class}
\name{Era5var-class}
\alias{Era5var-class}
\alias{Era5var}
\title{Classe Era5var pour extraire des donnees ERA5}
\arguments{
\item{area}{Nom de la zone.}

\item{coordinates}{Coordonnees delimitant la zone.}

\item{years}{Intervalle d'annees pour l'extraction.}

\item{variable_name}{Nom de la variable a extraire.}

\item{monthly}{(booleen, optionnel) Si TRUE, extrait les donnees mois par mois. Par defaut, FALSE.}

\item{directory}{Repertoire contenant les fichiers NetCDF.}
}
\value{
Un nouvel objet Era5var.
}
\description{
Cette classe contient des methodes pour extraire des donnees ERA5.
}
\section{Fields}{

\describe{
\item{\code{area_name}}{Nom de la zone pour laquelle les donnees sont necessaires (par exemple, 'France', 'London').}

\item{\code{coordinates}}{Liste des coordonnees nord, ouest, sud et est qui delimitent la zone.}

\item{\code{years_included}}{Intervalle d'annees pour l'extraction sous forme de chaîne (par exemple, '1961-1963') ou une seule annee (par exemple, '1961').}

\item{\code{variable_name}}{Nom de la variable a extraire (par exemple, 'total_precipitation').}

\item{\code{monthly}}{Booleen, si TRUE, l'extraction se fait mois par mois puis fusionnee (utile pour les grandes donnees).
Initialiser un objet Era5var avec les parametres specifies.}
}}


