% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{feature_set_default}
\alias{feature_set_default}
\title{Default feature names for the event classifiers}
\usage{
feature_set_default(kind = c("pathways", "topographic"))
}
\arguments{
\item{kind}{Either \code{"pathways"} or \code{"topographic"}.}
}
\value{
Character vector of feature names.
}
\description{
Request the default feature names used by the event classifiers.
}
\examples{
feature_set_default("pathways")

}
\seealso{
\code{\link[=features_pathways]{features_pathways()}}, \code{\link[=features_topographic]{features_topographic()}}, \code{\link[=classify_events]{classify_events()}}
}
