% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_barabasi.R
\name{gas}
\alias{gas}
\title{Assign preference scores}
\usage{
gas(mat, attribute.mat, intercept = 1, spatial.dist, beta)
}
\arguments{
\item{mat}{Network matrix.}

\item{attribute.mat}{Attribute matrix.}

\item{intercept}{Intercept.}

\item{spatial.dist}{Spatial distance among nodes.}

\item{beta}{Coefficients.}
}
\description{
Assign preference scores
}
