% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas.X.R
\name{gas.X}
\alias{gas.X}
\title{Assign attraction score}
\usage{
gas.X(mat, attribute.mat, base.prob = 0.05, spatial.dist)
}
\arguments{
\item{mat}{A network matrix.}

\item{attribute.mat}{A matrix for belief score (network edges).}

\item{base.prob}{Base probability. the default is .05.}

\item{spatial.dist}{Spatial distance between nodes.}
}
\description{
Assign attraction score
}
