% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas.X2.R
\name{gas.X2}
\alias{gas.X2}
\title{Assign attraction score 2}
\usage{
gas.X2(mat, attribute.mat, base.prob = 0.05, spatial.dist)
}
\arguments{
\item{mat}{A network matrix.}

\item{attribute.mat}{A matrix for belief score (network edges).}

\item{base.prob}{Base probability. the default is .05.}

\item{spatial.dist}{Spatial distance between nodes.}
}
\description{
Assign attraction score 2
}
