% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gas.R
\name{give.attra.score}
\alias{give.attra.score}
\title{Attribute Attractive Scores}
\usage{
give.attra.score(beta = 0.05)
}
\arguments{
\item{beta}{baseline probability. Default is set to as 0.05.}
}
\description{
Attribute attractive scores between two nodes to
decide how likely they are going to form a tie.
}
