% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate.matrix.R
\name{initiate.matrix}
\alias{initiate.matrix}
\title{Form an initial matrix for simulation}
\usage{
initiate.matrix(n)
}
\arguments{
\item{n}{the number of nodes.}
}
\description{
Form an initial matrix for simulation
}
