% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeaLevelData.R
\name{load_PSMSLdata}
\alias{load_PSMSLdata}
\title{Charger le fichier CSV contenant les donnees PSMSL}
\usage{
load_PSMSLdata(csv_path)
}
\arguments{
\item{csv_path}{Chemin vers le fichier CSV a charger.}
}
\value{
Affiche un message indiquant si le DataFrame a ete charge avec succes
ou si une erreur s'est produite.
}
\description{
Verifie si un fichier CSV specifique existe dans le repertoire de donnees.
Si le fichier existe, il est charge dans un DataFrame. Sinon, une erreur est levee.
}
\examples{
# Chemin du fichier CSV:
#csv_path <- './inst/externaldata/psmsl_data.csv'
#psmsl <- load_PSMSLdata(csv_path)

}
