% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinom.summary.R
\name{multinom.summary}
\alias{multinom.summary}
\title{Summary of multinational object}
\usage{
multinom.summary(out, rounding = TRUE, digits = 2, se.in.parenthesis = FALSE)
}
\arguments{
\item{out}{An output object of multinom in nnet library.}

\item{rounding}{If the result is rounded (Default: TRUE).}

\item{digits}{digit used for rounding.}

\item{se.in.parenthesis}{If the standard error is shown in parenthesis.}
}
\value{
Reference Reference category in dependent variable.

coef Result with the regression coefficients, standard errors and the p-value indicator.

fit  Fit measures of the regression.
}
\description{
Summary of multinational object
}
