% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisum.R
\name{multisum}
\alias{multisum}
\title{summary of multiple simulation results}
\usage{
multisum(list)
}
\arguments{
\item{list}{A list of simulated ACI objects.}
}
\value{
An evaluation dataframe with ACI parameters and variables used for simulation.
}
\description{
summary of multiple simulation results
}
