% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_para&des.R
\name{netdesc}
\alias{netdesc}
\title{Description regarding simulated network.}
\usage{
netdesc(sim.out)
}
\arguments{
\item{sim.out}{The simulated network.}
}
\description{
Description regarding simulated network.
}
