% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_para&des.R
\name{netsim.single}
\alias{netsim.single}
\title{Generate multiple random networks}
\usage{
netsim.single(
  mat.init,
  belief = 0,
  spatial.coord = NULL,
  target.outdegree = NULL,
  target.density = NULL,
  trial.limit = NULL,
  max.time = 10000,
  intercept = 1,
  homoph = 0,
  recip = 0,
  popul = 0,
  activ = 0,
  transiv = 0,
  common.target = 0,
  common.source = 0,
  proximity = 0,
  display.time = FALSE,
  seed = NULL,
  sparseM.output = FALSE,
  cores = 2,
  n.sim = 1
)
}
\arguments{
\item{mat.init}{Initial network matrix.}

\item{belief}{Belief scores.}

\item{spatial.coord}{Spatial coordination.}

\item{target.outdegree}{Target outdegree to end simulation.}

\item{target.density}{Target density to end simulation.}

\item{trial.limit}{The upper limits for trials.}

\item{max.time}{The maximum time to simulate one network.}

\item{intercept}{The intercept for prefernece score. Default is 1.}

\item{homoph}{Homophily beta.}

\item{recip}{Reciprocity beta.}

\item{popul}{Popularity beta.}

\item{activ}{Activity beta.}

\item{transiv}{Transitivity beta.}

\item{common.target}{Common target beta.}

\item{common.source}{Common source beta.}

\item{proximity}{Proximity beta.}

\item{display.time}{Display time. Default is *FALSE*.}

\item{seed}{Seed for random simulation.}

\item{sparseM.output}{Export sparse matrix. Default is *FALSE*.}

\item{cores}{The number of cores for parallel comuptation. Default is the *maximum-1*.}

\item{n.sim}{Number of simulation.}
}
\description{
Generate multiple random networks
}
