% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_interface.R
\name{nobs.aci_fit}
\alias{nobs.aci_fit}
\title{Number of observations behind a fitted latent-cause model}
\usage{
\method{nobs}{aci_fit}(object, ...)
}
\arguments{
\item{object}{An \code{aci_fit} object.}

\item{...}{Ignored, for consistency with the generic.}
}
\value{
1-length integer, the number of observations used in fitting.
}
\description{
Number of observations behind a fitted latent-cause model
}
