% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmulti.R
\name{plotmulti}
\alias{plotmulti}
\title{Visual comparison among multiple simulation}
\usage{
plotmulti(eval_dat, x, xlab)
}
\arguments{
\item{eval_dat}{evaluation data}

\item{x}{x index}

\item{xlab}{xlab}
}
\description{
Visual comparison among multiple simulation
}
