% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.obs_traj}
\alias{print.obs_traj}
\title{Print an observed trajectory}
\usage{
\method{print}{obs_traj}(x, ...)
}
\arguments{
\item{x}{An \code{obs_traj} object.}

\item{...}{Ignored, for consistency with the generic.}
}
\value{
\code{x}, invisibly; called for the printed output.
}
\description{
Print an observed trajectory
}
