% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_barabasi.R
\name{select.agent.altcandid}
\alias{select.agent.altcandid}
\title{Select Agents for ABM_Barabasi}
\usage{
select.agent.altcandid(target.outdegree, mat, actID)
}
\arguments{
\item{target.outdegree}{Target outdegree.}

\item{mat}{A network matrix.}

\item{actID}{ID.}
}
\description{
Select Agents for ABM_Barabasi
}
