% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.agent.R
\name{select.agent}
\alias{select.agent}
\title{Randomly select a pair of ego and alter for network simulation}
\usage{
select.agent(target.outdegree, mat, actID)
}
\arguments{
\item{target.outdegree}{The outdegree for selected ego.}

\item{mat}{A network matrix.}

\item{actID}{the ID of actors.}
}
\description{
Randomly select a pair of ego and alter for network simulation
}
