% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.aci.multi.R
\name{sim.aci.multi}
\alias{sim.aci.multi}
\title{ACI simulation for multiple parameters at once.}
\usage{
sim.aci.multi(
  n.sim,
  n,
  Belief,
  Degree,
  base.prob,
  homoph = 0,
  recip = 0,
  popul = 0,
  activ = 0,
  transiv = 0,
  common.target = 0,
  common.source = 0,
  display.time = FALSE,
  display.plot = FALSE,
  display.sim.time = FALSE
)
}
\arguments{
\item{display.sim.time}{}
}
\description{
ACI simulation for multiple parameters at once.
}
