% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.aci.para.R
\name{sim.aci.para}
\alias{sim.aci.para}
\title{Agent based Network simulation using multicore}
\usage{
sim.aci.para(
  n.sim,
  n,
  Belief,
  Degree,
  base.prob,
  core,
  homoph = 0,
  recip = 0,
  popul = 0,
  activ = 0,
  transiv = 0,
  common.target = 0,
  common.source = 0,
  display.time = FALSE,
  plotfamily = NA
)
}
\arguments{
\item{n.sim}{Number of simulation.}

\item{n}{Number of nodes.}

\item{Belief}{see \code{give.attra.score}.}

\item{Degree}{see \code{give.attra.score}.}

\item{base.prob}{see \code{give.attra.score}.}

\item{core}{The number of core for simulation.}

\item{homoph}{see \code{give.attra.score}.}

\item{recip}{see \code{give.attra.score}.}

\item{popul}{see \code{give.attra.score}.}

\item{activ}{see \code{give.attra.score}.}

\item{transiv}{see \code{give.attra.score}.}

\item{common.target}{see \code{give.attra.score}.}

\item{common.source}{see \code{give.attra.score}.}

\item{display.time}{see \code{give.attra.score}.}

\item{plotfamily}{Define a family of fonts for plot.}
}
\description{
Agent based Network simulation using multicore
}
