% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.aci.para2.R
\name{sim.aci.para2}
\alias{sim.aci.para2}
\title{ACI simulation test 2}
\usage{
sim.aci.para2(
  n.sim,
  core,
  seed = NULL,
  n,
  belief = 0,
  spatial.coord = NULL,
  target.outdegree = NULL,
  target.density = NULL,
  trial.limit = NULL,
  max.time = 10000,
  base.prob = 0.05,
  homoph = 0,
  recip = 0,
  popul = 0,
  activ = 0,
  transiv = 0,
  common.target = 0,
  common.source = 0,
  proximity = 0,
  display.time = FALSE,
  display.plot = FALSE,
  sparseM.output = FALSE
)
}
\arguments{
\item{n.sim}{Number of simulation.}

\item{sparseM.output}{}
}
\description{
ACI simulation test 2
}
