% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.aci.R, R/simulate_multi.R
\name{sim.aci}
\alias{sim.aci}
\title{ACI simulation: single}
\usage{
sim.aci(nsim, actor, belief, degree, plot = FALSE, xlab)

sim.aci(nsim, actor, belief, degree, plot = FALSE, xlab)
}
\arguments{
\item{nsim}{The number of simulation.}

\item{actor}{The number of nodes.}

\item{belief}{The belief score.}

\item{degree}{A \code{numeric} object of average degree used or
a \code{list} object of multiple average degrees used.}

\item{plot}{Only applies when \code{degree} is a \code{list} object.}

\item{xlab}{Assign \code{ave.deg} or \code{density} as xlab when \code{plot = TRUE}.}

\item{n.sim}{Number of simulation.}

\item{n}{Number of nodes.}

\item{Belief}{see \code{give.attra.score}.}

\item{Degree}{see \code{give.attra.score}.}

\item{base.prob}{see \code{give.attra.score}.}

\item{homoph}{see \code{give.attra.score}.}

\item{recip}{see \code{give.attra.score}.}

\item{popul}{see \code{give.attra.score}.}

\item{activ}{see \code{give.attra.score}.}

\item{transiv}{see \code{give.attra.score}.}

\item{common.target}{see \code{give.attra.score}.}

\item{common.source}{see \code{give.attra.score}.}

\item{display.time}{see \code{give.attra.score}.}

\item{display.plot}{see \code{give.attra.score}.}

\item{display.sim.time}{see \code{give.attra.score}.}
}
\description{
ACI simulation: single

Simulation: ACI simulation based on random level of average degree.
}
\author{
Hao Liang \url{hl893@cornell.edu}
Keiichi Satoh \url{keiichi.satoh@r.hit-u.ac.jp}
}
