% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multi.R
\name{sim.degree}
\alias{sim.degree}
\title{Simulation: ACI simulation based on random level of average degree.}
\usage{
sim.degree(nsim, actor, belief, degree, plot = FALSE)
}
\arguments{
\item{nsim}{The number of simulation.}

\item{actor}{The number of nodes.}

\item{belief}{The belief score.}

\item{degree}{A \code{numeric} object of average degree used or
a \code{list} object of multiple average degrees used.}

\item{plot}{Only applies when \code{degree} is a \code{list} object.}
}
\description{
Simulation: ACI simulation based on random level of average degree.
}
\author{
Hao Liang \url{hl893@cornell.edu}
Keiichi Satoh \url{keiichi.satoh@r.hit-u.ac.jp}
}
