% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeaLevelData.R
\docType{data}
\name{source_dir}
\alias{source_dir}
\title{Copier et renommer les fichiers en fonction de l'abréviation du pays}
\format{
An object of class \code{character} of length 1.
}
\usage{
source_dir
}
\arguments{
\item{abbreviation}{Abréviation du pays pour filtrer et copier les fichiers associés}

\item{source_dir}{Répertoire source contenant les fichiers .rlrdata}

\item{target_dir}{Répertoire cible où les fichiers seront copiés et renommés}

\item{df}{DataFrame contenant les données PSMSL pour filtrer par abréviation de pays}
}
\value{
NULL Renvoie NULL si aucune entrée n'est trouvée pour l'abréviation du pays
}
\description{
Copie les fichiers .rlrdata d'un répertoire source vers un répertoire cible basé
sur l'abréviation du pays fournie. Les fichiers copiés sont ensuite renommés en .txt.
}
\examples{
# Exécution de la fonction pour une abréviation spécifique
copy_and_rename_files_by_country("FRA")
# (Voir le code complet ci-dessous pour l'exécution)

}
\keyword{datasets}
