% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeaLevelData.R
\docType{data}
\name{url}
\alias{url}
\title{Télécharge et décompresse les données RLR mensuelles}
\format{
An object of class \code{character} of length 1.
}
\usage{
url
}
\arguments{
\item{url}{URL du fichier à télécharger}

\item{destination_dir}{Chemin vers le répertoire de destination}

\item{zip_file_path}{Chemin complet vers le fichier ZIP}

\item{extract_path}{Chemin complet vers le répertoire d'extraction}
}
\value{
Affiche un message indiquant si le téléchargement et la décompression sont terminés,
ou si le dossier existe déjà
}
\description{
Télécharge un fichier ZIP contenant les données RLR mensuelles depuis une URL spécifiée,
le décompresse dans un répertoire de destination, et supprime le fichier ZIP après l'extraction.
Si le dossier cible existe déjà, aucune action n'est effectuée.
}
\examples{
# Exemple de chemin d'accès et d'utilisation
url <- "https://psmsl.org/data/obtaining/rlr.monthly.data/rlr_monthly.zip"
destination_dir <- "../data/required_data"
zip_file_path <- file.path(destination_dir, "rlr_monthly.zip")
extract_path <- file.path(destination_dir, "rlr_monthly")
# Exécute le téléchargement et l'extraction
# (Voir le code complet ci-dessous pour l'exécution)

}
\keyword{datasets}
