% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windcomponent.R
\docType{class}
\name{WindComponent-class}
\alias{WindComponent-class}
\alias{WindComponent}
\title{WindComponent Class}
\arguments{
\item{u10_file}{Character. File path for the u10 component of wind.}

\item{v10_file}{Character. File path for the v10 component of wind.}

\item{mask_file}{Character. File path for the mask file.}

\item{dt}{Data.table. The data table from which duplicates are to be removed.}

\item{lat}{Numeric. Latitude of the location to plot.}

\item{lon}{Numeric. Longitude of the location to plot.}

\item{reference_period}{Character vector. Reference period for calculating exceedance frequency.}
}
\value{
Data.table. Data table with duplicates removed.

Calculate wind power

Data.table. Wind power data table.
Calculate wind power thresholds

Data.table. Wind power thresholds data table.

Calculate days above wind power thresholds

Data.table. Days above threshold data table.

Calculate wind exceedance frequency

Data.table. Wind exceedance frequency data table.

Calculate standardized wind exceedance frequency

Data.table. Standardized wind exceedance frequency data table.

Plot wind exceedance frequency

ggplot object. Plot of wind exceedance frequency for the specified location.
Plot wind exceedance frequency map

ggplot object. Map of wind exceedance frequency.
}
\description{
A class for analyzing wind data from NetCDF files, including calculating wind power, wind thresholds, days above thresholds, wind exceedance frequency, and plotting results.
}
\section{Fields}{

\describe{
\item{\code{u10_file}}{Character. File path for the u10 component of wind.}

\item{\code{v10_file}}{Character. File path for the v10 component of wind.}

\item{\code{mask_file}}{Character. File path for the mask file.}

\item{\code{reference_period}}{Character. Reference period for calculating statistics.}

\item{\code{u10_dt}}{Data.table. Data table for u10 component data.}

\item{\code{v10_dt}}{Data.table. Data table for v10 component data.}

\item{\code{mask_dt}}{Data.table. Data table for mask data.}

\item{\code{wind_power_dt}}{Data.table. Data table for calculated wind power.}

\item{\code{wind_thresholds_dt}}{Data.table. Data table for calculated wind thresholds.
Initialize WindComponent object}
}}


