% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACI-package.R
\docType{package}
\name{ACI-package}
\alias{ACI}
\alias{ACI-package}
\title{ACI: Package ACI for Actuarial Climate Index}
\description{
This package implements several methods that enable to compute the Actuarial Climate Index on different regions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/XavierMilhaud/ACI-Rpackage}
  \item Report bugs at \url{https://github.com/XavierMilhaud/ACI-Rpackage/issues}
}

}
\author{
\strong{Maintainer}: Xavier Milhaud \email{xavier.milhaud.research@gmail.com} [contributor]

Authors:
\itemize{
  \item Dabakh Kane \email{kanedabakh55@gmail.com} [contributor]
}

Other contributors:
\itemize{
  \item Esteban Mauboussin \email{esteban.mauboussin@cnp.fr} [contributor]
}

}
\keyword{internal}
