% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{aci}
\alias{aci}
\title{Calculating Advocacy Coalition Index}
\usage{
aci(matrix, policy.score, alpha = 0.5, print = FALSE)
}
\arguments{
\item{matrix}{An advocacy coalition network matrix.}

\item{policy.score}{A numeric object of belief scores.}

\item{alpha}{threshold uses to decide the cut off value.}

\item{print}{whether to print subsystem-level ACI output in Console. Default is \code{FALSE}.}
}
\value{
Advocacy Coalition Index in three levels:
    Subsystem-level, Coalition-level, and actor-level.
}
\description{
Calculating Advocacy Coalition Index
}
