% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActuarialClimateIndex.R
\docType{class}
\name{ActuarialClimateIndex-class}
\alias{ActuarialClimateIndex-class}
\alias{ActuarialClimateIndex}
\title{Classe ActuarialClimateIndex}
\arguments{
\item{temperature_data_path}{Chemin vers le fichier de données de température}

\item{precipitation_data_path}{Chemin vers le fichier de données de précipitation}

\item{wind_u10_data_path}{Chemin vers le fichier de données du composant U du vent}

\item{wind_v10_data_path}{Chemin vers le fichier de données du composant V du vent}

\item{country_abbrev}{Abréviation du pays pour le niveau de la mer}

\item{mask_data_path}{Chemin vers le fichier de masque géographique}

\item{study_period}{Période d'étude pour l'analyse (vecteur de deux dates)}

\item{reference_period}{Période de référence pour la standardisation (vecteur de deux dates)}

\item{factor}{Facteur d'ajustement pour le niveau de la mer}
}
\value{
Un nouvel objet ActuarialClimateIndex
Calculate Actuarial Climate Index

Cette méthode calcule l'indice climatique actuariel (ACI) en utilisant les données des différents composants climatiques initialisés.

Un data.table contenant les valeurs calculées de l'indice climatique actuariel (ACI)
}
\description{
Cette classe représente l'indice climatique actuariel, qui est calculé à partir de plusieurs composants climatiques tels que la température, les précipitations, la sécheresse, le vent, et le niveau de la mer.
}
\section{Fields}{

\describe{
\item{\code{temperature_component}}{Composant température de type ANY}

\item{\code{precipitation_component}}{Composant précipitation de type ANY}

\item{\code{drought_component}}{Composant sécheresse de type ANY}

\item{\code{wind_component}}{Composant vent de type ANY}

\item{\code{sealevel_component}}{Composant niveau de la mer de type ANY}

\item{\code{study_period}}{Période d'étude pour l'analyse (vecteur de deux dates)}

\item{\code{reference_period}}{Période de référence pour la standardisation (vecteur de deux dates)
Initialize ActuarialClimateIndex object

Cette méthode initialise un objet de la classe ActuarialClimateIndex avec les chemins des fichiers de données nécessaires et les périodes d'étude et de référence.}
}}


