% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.network.R
\name{create.network}
\alias{create.network}
\title{Simulation: Create Simulating Networks}
\usage{
create.network(
  n,
  Belief,
  Degree,
  base.prob = 0.05,
  homoph = 0,
  recip = 0,
  popul = 0,
  activ = 0,
  transiv = 0,
  common.target = 0,
  common.source = 0,
  display.time = FALSE,
  display.plot = FALSE,
  plotfamily = NA,
  max.time = 1e+05
)
}
\arguments{
\item{n}{The number of nodes.}

\item{Belief}{The belief scores of each nodes.}

\item{Degree}{The target degree.}

\item{base.prob}{see \code{give.attra.score}.}

\item{homoph}{see \code{give.attra.score}.}

\item{recip}{see \code{give.attra.score}.}

\item{popul}{see \code{give.attra.score}.}

\item{activ}{see \code{give.attra.score}.}

\item{transiv}{see \code{give.attra.score}.}

\item{common.target}{see \code{give.attra.score}.}

\item{common.source}{see \code{give.attra.score}.}

\item{display.time}{display time for simulation, default is \code{False}.}

\item{display.plot}{display simulation results in plot, default is \code{False}.}

\item{plotfamily}{Define a family of fonts for plot.}

\item{max.time}{maximum time for simulation, default is \code{100000}.}
}
\value{
\code{time}: time used for creating the network,
        \code{mat}: the imported network matrix,
        \code{Belief}: the belief scores of each nodes,
        \code{Degree}: the target degree.
        \code{Record}: the record of edge formation.
}
\description{
Network simulation based on agent-based models (ABM).
}
