% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multi.R
\name{multisum}
\alias{multisum}
\title{Summary}
\usage{
multisum(list, x, xlab)
}
\arguments{
\item{list}{a list of object}

\item{x}{see \code{plotmulti}.}

\item{xlab}{see \code{plotmulti}.}
}
\description{
Summary
}
