% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmulti.R
\name{plotmulti}
\alias{plotmulti}
\title{Visual comparison among multiple simulation}
\usage{
plotmulti(eval_dat)
}
\arguments{
\item{eval_dat}{evaluation data}
}
\description{
Visual comparison among multiple simulation
}
