% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmsl.R
\docType{data}
\name{psmsl}
\alias{psmsl}
\title{PSMSL datasets on Sea Levels}
\format{
\subsection{\code{psmsl}}{

A data table (and thus also data frame) with 1,585 rows and 9 columns:
\describe{
\item{Station Name}{(Character) The station name}
\item{ID}{(Integer) The raw id}
\item{Lat.}{(Numeric) The latitude of the station}
\item{Lon.}{(Numeric) The longitude of the station}
\item{GLOSS ID}{(Integer) The id in the gloassary}
\item{Country}{(Character) The country of the station}
\item{Date}{(Character) The date of measurement}
\item{Coastline}{(Character) The code of the coastline}
\item{Station}{(Character) The station id}
}
}
}
\source{
\url{https://psmsl.org/data/obtaining/}
}
\usage{
psmsl
}
\description{
An example of datasets extracted from website giving...
}
\keyword{datasets}
