% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.agent.R
\name{select.agent}
\alias{select.agent}
\title{Simulation: Select agent at random}
\usage{
select.agent(Act.ID = NULL, Target.Degree, mat)
}
\arguments{
\item{Act.ID}{An ID list of actors.}

\item{Target.Degree}{The number of degree.}

\item{mat}{A network matrix.}
}
\description{
Simulation: Select agent at random
}
