% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.agent.R
\name{select.agent}
\alias{select.agent}
\title{Simulation: Select agent at random}
\usage{
select.agent(Target.Degree, mat, Act.ID)
}
\arguments{
\item{Target.Degree}{The number of degree.}

\item{mat}{A network matrix.}

\item{Act.ID}{An ID list of actors.}
}
\description{
Simulation: Select agent at random
}
