% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.aci.R
\name{sim.aci}
\alias{sim.aci}
\title{ACI simulation: single}
\usage{
sim.aci(
  n.sim,
  n,
  Belief,
  Degree,
  base.prob,
  homoph = 0,
  recip = 0,
  popul = 0,
  activ = 0,
  transiv = 0,
  common.target = 0,
  common.source = 0,
  display.time = FALSE,
  display.plot = FALSE,
  display.sim.time = FALSE
)
}
\arguments{
\item{n.sim}{Number of simulation.}

\item{n}{Number of nodes.}

\item{Belief}{see \code{give.attra.score}.}

\item{Degree}{see \code{give.attra.score}.}

\item{base.prob}{see \code{give.attra.score}.}

\item{homoph}{see \code{give.attra.score}.}

\item{recip}{see \code{give.attra.score}.}

\item{popul}{see \code{give.attra.score}.}

\item{activ}{see \code{give.attra.score}.}

\item{transiv}{see \code{give.attra.score}.}

\item{common.target}{see \code{give.attra.score}.}

\item{common.source}{see \code{give.attra.score}.}

\item{display.time}{see \code{give.attra.score}.}

\item{display.plot}{see \code{give.attra.score}.}

\item{display.sim.time}{see \code{give.attra.score}.}
}
\description{
ACI simulation: single
}
